sm6150-common: Enable Gboard side padding
This adds padding to the left and right sides of the keyboard when using
Gboard.

In portrait mode, there is 4dp of padding to align the keyboard with
some UI elements and separate the keys from the edge of the screen.

In landscape mode, there is 64dp of padding to reduce the overall key
width, which makes the keyboard significantly easier to use.

TODO: It might be better to set these in individual device trees, but
for now, these props seem to work well for all official devices.

Suggested-by: Diab Neiroukh <lazerl0rd@thezest.dev>
Change-Id: I3cc44d342a11b5b25130eed61470e8d98c6a5d83
Signed-off-by: Cykeek <Cykeek@proton.me>